Commit related to Interfaces
An introduction to Interfaces is given in this program. A very beautiful concept of applying restrictions to interfaces is seen. When dealing with 100s of lines of code, it is not possible to check if an unwanted method (something not present in the interface) is implemented by a class. There is no way of forcing a developer to not do such an evil thing when it comes to a hypothetical file consisting of say a million lines. But it would be able to add restriction to it, if such a thing is present by using the concept of dynamic binding, and this is shown in this demonstration.